requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.1
streamlit>=1.40.0
chromadb>=0.5.0
beautifulsoup4>=4.12.3
tiktoken>=0.7.0
